Check Permissions Before Deleting Messages

+++ src/ClearMessageListener.ts
@@ -8,8 +8,17 @@ export class ClearMessageListener implements MessageListener {
     }
 
     public process(msg: Message): Promise<any> {
+
+        const howMany = parseInt(msg.content.substring(6), 10) || 5;
+
+        if (!msg.guild.member(msg.author).hasPermission("clear")) {
+            msg.author.send("clearing messages is prohibited for you!");
+            return Promise.reject(new Error(`User '${msg.author.username}' attempted to clear \
+                messages on channel ${msg.channel.toString()}`));
+        }
+
         return msg.channel.fetchMessages({
-            limit: 100,
+            limit: howMany,
         }).then((messages: Collection<Snowflake, Message>) => {
             return messages.map((message) => message.delete());
         });
